+++ android/src/main/kotlin/com/ridehovr/hovr_app_update/AppUpdateChannelConstants.kt
@@ -5,4 +5,6 @@ internal object AppUpdateChannelConstants {
     const val DIALOG_TAG = "hovr_app_update_dialog"
     const val METHOD_CONFIGURE = "configure"
     const val METHOD_PROMPT = "promptIfUpdateRequired"
+    const val METHOD_GET_INSTALLED_VERSION = "getInstalledVersion"
+    const val METHOD_GET_APP_INFO = "getAppInfo"
 }