+++ lib/hovr_app_update.dart
@@ -1,4 +1,5 @@
 library;
 
+export 'src/app_info.dart';
 export 'src/app_update_config.dart';
 export 'src/hovr_app_update.dart';